requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
chainer==7.8.1
charset-normalizer==2.1.1
click==8.1.3
cnstd==1.2
colorama==0.4.6
coloredlogs==15.0.1
cssselect==1.2.0
cssutils==2.6.0
cvzone==1.5.6
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
dill==0.3.6
easyocr==1.6.2
et-xmlfile==1.1.0
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
Flask-Babel==2.0.0
flatbuffers==22.10.26
fonttools==4.28.2
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
humanfriendly==10.0
idna==3.4
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==5.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.3.2
lanms-neo==1.0.2
lmdb==1.3.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.0
mediapipe==0.8.11
mpmath==1.2.1
multidict==6.0.2
multiprocess==0.70.14
networkx==2.6.3
ninja==1.10.2.4
numpy==1.22.4
oauthlib==3.2.2
onnx==1.12.0
onnxruntime==1.13.1
opencv-contrib-python==4.5.4.60
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddleocr==2.6.1.0
paddlepaddle==2.3.2
pandas==1.5.1
pdf2docx==0.5.6
Pillow==9.2.0
Polygon3==3.0.9.1
premailer==3.10.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycodestyle==2.9.1
pycryptodome==3.15.0
pyDeprecate==0.3.2
PyMuPDF==1.19.0
pyparsing==3.0.6
pyreadline3==3.4.1
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
pytorch-lightning==1.7.7
pytz==2022.6
PyWavelets==1.2.0
PyYAML==6.0
rapidfuzz==2.13.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.3
scipy==1.7.3
setuptools-scm==6.3.2
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.3.2.post1
sympy==1.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchmetrics==0.10.2
torchvision==0.11.1
tqdm==4.64.1
typing_extensions==4.0.1
Unidecode==1.3.6
urllib3==1.26.12
visualdl==2.4.1
waifu2x==2022.2
Wand==0.6.10
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
